13: GET Validation refactoring: package refactoring
